requirements.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

Django==1.4.2

django-debug-toolbar==0.9.4
django-staticfiles==1.2.1
django_compressor==1.3

-e git+https://github.com/ntucker/django-user-accounts.git@2adaeefea133c41357a178bb5581a446d15852b5#egg=django-user-accounts
django-timezones==0.2
pytz==2012j
django-openid==0.3a1
python-openid==2.2.5

-e git+https://github.com/ntucker/django-announcements.git@6208103674f89942559827ba6fbf029671e7e9ab#egg=django-announcements
django-pagination==1.0.7
-e git+https://github.com/ntucker/django-notification.git@8ad545ae85ce1f41985aa824457844c7632cac04#egg=django-notification
-e git+https://github.com/ntucker/idios.git@07c6d21a681cd640259af6f3d2070588dc72aa04#egg=idios
metron==0.1

pinax-theme-bootstrap==2.0.1
-e git+https://github.com/pinax/pinax-theme-bootstrap-account.git@98386a90776e76200b9aa6eefe7e6a7b5e2ed553#egg=pinax-theme-boostrap-account

django-pure-pagination==0.2.1
pytils==0.2.3

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

gondor==1.2.1
psycopg2==2.4.5
gunicorn==0.17.2
eventlet==0.9.16
sqlalchemy==0.7.8
#lxml==2.3
pillow==2.3.0
html5lib==0.90
bleach==1.2.1
slimit==0.6
recaptcha-client==1.0.6
south==0.7.6
raven==2.0.3
-e git+https://github.com/ntucker/django-cache-magic.git@dbfb4aa042f4c5b1fb145bdb1d5f721a8909e2c8#egg=django-cache-magic
-e git+https://github.com/rdegges/django-clear-cache.git@9251fd3730afb8ab3deaf5c2619889438fc760c4#egg=django-clear-cache
johnny-cache==1.4

#celery and its dependencies
amqplib==1.0.2
anyjson==0.3.3
kombu==2.5.4
python-dateutil==2.1
billiard==2.7.3.18
celery==3.0.13
django-celery==3.0.11
tornado==2.4.1
celerymon==1.0.3

django-redis-cache==0.9.2
sorl-thumbnail==11.12
django-social-auth==0.6.1
-e git+https://github.com/arneb/django-messages.git@80316aa1b57f5dda987a2e580ec78372974a7520#egg=django-messages
django-tinymce==1.5.1b2
django-cms==2.3.8
phileo==1.0

# Need to specify direct path to postmarkup it seems.
https://postmarkup.googlecode.com/files/postmarkup-1.2.0.tar.gz
-e git+https://github.com/ntucker/pybbm.git@d1d0a3c92986fa012b01105677b38c35fe59d5cf#egg=pybbm

-e hg+https://bitbucket.org/ntucker/django-recaptcha/@2e55e507faa9e9c6c97b06b48c75641cc4df0e8b#egg=django-recaptcha

# following is needed for a migration, it is harder to remove the parts of the
# migration that it is to add this back.
django-imagekit

djangorestframework==2.3.14
django-cors-headers
django-filter
requests==2.3.0
